upgrade_deadlines/requirements.txt
edx-rest-api-client==1.7.1
python-dateutil==2.6.0
